+++ tests/smoke/core/__init__.py
@@ -1,0 +1,41 @@
+from smoke.core.cli import cmd
+from smoke.core.commons import (
+    Case,
+    Step,
+    WaitUntil,
+    write_cases,
+    Equal,
+    Empty,
+    NotEmpty,
+    True_,
+    False_,
+    IsInstance,
+    Len,
+    In,
+    Contains,
+)
+from smoke.core.settings import (
+    SmokeSettings,
+    get_settings,
+    get_rule_source,
+)
+
+__all__ = [
+    'Case',
+    'Step',
+    'WaitUntil',
+    'write_cases',
+    'Equal',
+    'Empty',
+    'NotEmpty',
+    'True_',
+    'False_',
+    'IsInstance',
+    'Len',
+    'In',
+    'Contains',
+    'SmokeSettings',
+    'get_settings',
+    'get_rule_source',
+    'cmd',
+]